Move getCatalogWithCriteriaId to a helpers file in state
